# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: VaultRun API
  version: "1.0.0"
  description: |
    REST API for the VaultRun sandbox execution platform.
    All `/api/v1/*` endpoints require an API key passed in the `X-API-Key` header.
    Key-management endpoints additionally require the master key.
  contact:
    name: VaultRun
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Pagination:
      type: object
      required: [page, limit, offset, total]
      properties:
        page:   { type: integer, example: 1 }
        limit:  { type: integer, example: 50 }
        offset: { type: integer, example: 0 }
        total:  { type: integer, example: 142 }

    # ── Sessions ───────────────────────────────────────────────────────────────
...
    File:
      type: object
      required: [id, session_id, path, size_bytes, content_type, created_at]
      properties:
        id:           { type: string, format: uuid }
        session_id:   { type: string, format: uuid }
        path:         { type: string, example: "/workspace/script.py" }
        size_bytes:   { type: integer }
        content_type: { type: string, example: "text/x-python" }
        created_at:   { type: string, format: date-time }

    # ── API Keys ───────────────────────────────────────────────────────────────
...
security:
  - ApiKeyAuth: []

# ──────────────────────────────────────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/sessions/{id}/files:
    post:
      tags: [files]
      summary: Upload a file to the session workspace
      parameters:
        - { $ref: '#/components/parameters/ResourceID' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, path]
              properties:
                file:
                  type: string
                  format: binary
                path:
                  type: string
                  description: "Destination path in the workspace, e.g. script.py"
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/File' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

    get:
      tags: [files]
      summary: List files in the session workspace
      parameters:
        - { $ref: '#/components/parameters/ResourceID' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items: { $ref: '#/components/schemas/File' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /api/v1/sessions/{id}/files/{path}:
    get:
      tags: [files]
      summary: Download a single file from the workspace
      parameters:
        - { $ref: '#/components/parameters/ResourceID' }
        - name: path
          in: path
          required: true
          schema: { type: string }
          description: "File path relative to the workspace root, e.g. output/result.json"
      responses:
        "200":
          description: File contents
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [files]
      summary: Delete a file from the workspace
      parameters:
        - { $ref: '#/components/parameters/ResourceID' }
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: File deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── Audit ──────────────────────────────────────────────────────────────────
...